Move chat registry dumps into nbt files (#3763)
